# openapi.yaml
openapi: 3.1.0
info:
  title: Colette API
  description: Public REST API for the Colette app. Supports OAuth 2.0 and API key authentication.
  license:
    name: MIT
  version: 0.1.0
servers:
- url: /api
paths:
...
  /subscriptions:
    get:
      tags:
      - Subscriptions
      description: List user subscriptions
      operationId: listSubscriptions
      parameters:
      - name: filterByTags
        in: query
        description: Whether to filter by tags linked to the subscription
        required: false
        schema:
          type: boolean
      - name: tag[]
        in: query
        description: Filter by the IDs of the tags linked to the subscription
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: withFeed
        in: query
        description: Whether to include the feed associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withUnreadCount
        in: query
        description: Whether to include the count of the unread subscription entries associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withTags
        in: query
        description: Whether to include the tags linked to the subscription
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Paginated list of subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_SubscriptionDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Subscriptions
      description: Subscribe to a web feed
      operationId: createSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreate'
        required: true
      responses:
        '201':
          description: Created subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Subscription already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /subscriptions/{id}:
    get:
      tags:
      - Subscriptions
      description: Get a subscription by ID
      operationId: getSubscription
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      - name: withFeed
        in: query
        description: Whether to include the feed associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withUnreadCount
        in: query
        description: Whether to include the count of the unread subscription entries associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withTags
        in: query
        description: Whether to include the tags linked to the subscription
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Subscription by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - Subscriptions
      description: Delete a subscription by ID
      operationId: deleteSubscription
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Successfully deleted subscription
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Subscriptions
      description: Update a subscription by ID
      operationId: updateSubscription
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionUpdate'
        required: true
      responses:
        '200':
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /subscriptions/{sid}/entries/{eid}/markAsRead:
    post:
      tags:
      - Subscriptions
      description: Mark a subscription entry as read
      operationId: markSubscriptionEntryAsRead
      parameters:
      - name: sid
        in: path
        description: Unique identifier of the associated subscription
        required: true
        schema:
          type: string
          format: uuid
      - name: eid
        in: path
        description: Unique identifier of the associated feed entry
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Updated subscription entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionEntry'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /subscriptions/{sid}/entries/{eid}/markAsUnread:
    post:
      tags:
      - Subscriptions
      description: Mark a subscription entry as unread
      operationId: markSubscriptionEntryAsUnread
      parameters:
      - name: sid
        in: path
        description: Unique identifier of the associated subscription
        required: true
        schema:
          type: string
          format: uuid
      - name: eid
        in: path
        description: Unique identifier of the associated feed entry
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Updated subscription entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionEntry'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /subscriptionEntries:
    get:
      tags:
      - Subscription Entries
      description: List subscription entries
      operationId: listSubscriptionEntries
      parameters:
      - name: streamId
        in: query
        description: Filter by the ID of a stream whose filters may apply to the subscription entry
        required: false
        schema:
          type: string
          format: uuid
      - name: subscriptionId
        in: query
        description: Filter by the ID of the associated subscription
        required: false
        schema:
          type: string
          format: uuid
      - name: hasRead
        in: query
        description: Filter by whether the subscription entry has been marked as read
        required: false
        schema:
          type: boolean
      - name: tag[]
        in: query
        description: Filter by the IDs of the tags linked to the associated subscription
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: cursor
        in: query
        description: Pagination cursor from the previous set of results
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Paginated list of subscription entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_SubscriptionEntryDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          type: string
    ApiErrorCode:
      type: string
      enum:
      - NOT_AUTHENTICATED
      - NOT_AUTHORIZED
      - NOT_FOUND
      - ALREADY_EXISTS
      - VALIDATION
      - BAD_GATEWAY
      - UNKNOWN
...
    Feed:
      type: object
      description: RSS feed
      required:
      - id
      - sourceUrl
      - link
      - title
      - description
      - refreshedAt
      - isCustom
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the feed
        sourceUrl:
          type: string
          format: uri
          description: URL to scrape for feed updates
        link:
          type: string
          format: uri
          description: URL of the webpage the feed links to
        title:
          type: string
          description: Title of the feed
        description:
          type:
          - string
          - 'null'
          description: Description of the feed
        refreshedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the feed was refreshed
        isCustom:
          type: boolean
          description: Whether the feed was scraped from a custom plugin
...
    FeedEntry:
      type: object
      description: RSS feed entry
      required:
      - id
      - link
      - title
      - publishedAt
      - description
      - author
      - thumbnailUrl
      - feedId
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the feed entry
        link:
          type: string
          format: uri
          description: URL of the webpage the feed entry links to
        title:
          type: string
          description: Title of the feed entry
        publishedAt:
          type: string
          format: date-time
          description: Timestamp at which the feed entry was published
        description:
          type:
          - string
          - 'null'
          description: Description of the feed entry
        author:
          type:
          - string
          - 'null'
          description: Author of the feed entry
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the feed entry
        feedId:
          type: string
          format: uuid
          description: Unique identifier of the associated RSS feed
...
    Paginated_SubscriptionDetails:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: Extended details of a user subscription
            required:
            - subscription
            properties:
              subscription:
                $ref: '#/components/schemas/Subscription'
                description: Subscription itself, always present
              feed:
                $ref: '#/components/schemas/Feed'
                description: Associated RSS feed, present if requested
              tags:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
                description: Linked tags, present if requested
              unreadCount:
                type: integer
                format: int64
                description: Count of unread subscription entries associated with the subscription, present if requested
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
    Paginated_SubscriptionEntryDetails:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: Extended details of a subscription entry
            required:
            - subscriptionEntry
            properties:
              subscriptionEntry:
                $ref: '#/components/schemas/SubscriptionEntry'
                description: Subscription entry itself, always present
              feedEntry:
                $ref: '#/components/schemas/FeedEntry'
                description: Associated feed entry, present if requested
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
...
    Subscription:
      type: object
      description: User subscription to an RSS feed
      required:
      - id
      - title
      - description
      - feedId
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the subscription
        title:
          type: string
          description: Human-readable name of the subscription
        description:
          type:
          - string
          - 'null'
          description: Description of the subscription
        feedId:
          type: string
          format: uuid
          description: Unique identifier of the associated RSS feed
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the subscription was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the subscription was modified
    SubscriptionCreate:
      type: object
      description: Data to create a new user subscription
      required:
      - title
      - feedId
      properties:
        title:
          type: string
          description: Human-readable name for the new subscription, cannot be empty
          minLength: 1
        description:
          type:
          - string
          - 'null'
          description: Description for the new subscription, cannot be empty
          minLength: 1
        feedId:
          type: string
          format: uuid
          description: Unique identifier of the associated RSS feed
    SubscriptionDetails:
      type: object
      description: Extended details of a user subscription
      required:
      - subscription
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
          description: Subscription itself, always present
        feed:
          $ref: '#/components/schemas/Feed'
          description: Associated RSS feed, present if requested
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Linked tags, present if requested
        unreadCount:
          type: integer
          format: int64
          description: Count of unread subscription entries associated with the subscription, present if requested
    SubscriptionEntry:
      type: object
      description: Association of a RSS feed entry to a user subscription. The pairing of subscription ID and feed entry ID is unique.
      required:
      - subscriptionId
      - feedEntryId
      - hasRead
      properties:
        subscriptionId:
          type: string
          format: uuid
          description: Unique identifier of the associated subscription
        feedEntryId:
          type: string
          format: uuid
          description: Unique identifier of the associated feed entry
        hasRead:
          type: boolean
          description: Whether the subscription entry has been marked as read
        readAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the subscription entry has been marked as read
...
    SubscriptionUpdate:
      type: object
      description: Updates to make to an existing subscription
      properties:
        title:
          type:
          - string
          - 'null'
          description: Human-readable name for the subscription to update, cannot be empty
          minLength: 1
        description:
          type:
          - string
          - 'null'
          description: Description for the subscription to update, cannot be empty
          minLength: 1
    Tag:
      type: object
      description: Tag that can be attached to subscriptions and bookmarks
      required:
      - id
      - title
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the tag
        title:
          type: string
          description: Human-readable name of the tag, unique per user
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was last modified
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []